Introduce parent Project class in entities
