+++ src/test/java/todoist/entities/Project.java
@@ -1,0 +1,15 @@
+package todoist.entities;
+
+public class Project {
+
+    private String name;
+
+    public String getName() {
+        return this.name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
