+++ src/test/java/todoist/entities/ProjectRequest.java
@@ -1,16 +1,6 @@
 package todoist.entities;
 
-public class ProjectRequest {
-
-    private String name;
-
-    public String getName() {
-        return this.name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
+public class ProjectRequest extends Project {
 
     public ProjectRequest(String name) {
         this.setName(name);
