+++ src/test/java/todoist/entities/ProjectResponse.java
@@ -3,10 +3,11 @@ package todoist.entities;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 // TODO: Investigate @JsonPropertyOrder, looks like it specifies the expected order
-public class ProjectResponse {
+public class ProjectResponse extends Project {
 
     private int id;
     private int order;
+    // TODO: Remove
     private String name;
     private int commentCount;
 
@@ -18,6 +19,7 @@ public class ProjectResponse {
         return order;
     }
 
+    // TODO: Remove next 2 methods
     public String getName() {
         return name;
     }
